% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addresses.R
\name{addressesBalance}
\alias{addressesBalance}
\title{Address Balance}
\usage{
addressesBalance(address, token_id = NULL)
}
\arguments{
\item{address}{Required: the P2PK address.}

\item{token_id}{Token ID.}
}
\value{

}
\description{
Current $ERG or token balance of an address.
}
