% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addresses.R
\name{addressesBalanceHeight}
\alias{addressesBalanceHeight}
\title{Address Balance At Height}
\usage{
addressesBalanceHeight(address, height, token_id = NULL)
}
\arguments{
\item{address}{Required: the P2PK address.}

\item{height}{Required: block height.}

\item{token_id}{Token ID.}
}
\value{

}
\description{
$ERG or token balancee of an address at a specific block height.
}
