% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addresses.R
\name{addressesBalanceHistory}
\alias{addressesBalanceHistory}
\title{Address Balance History}
\usage{
addressesBalanceHistory(address, token_id = NULL)
}
\arguments{
\item{address}{Required: the P2PK address.}

\item{token_id}{Token ID.}
}
\value{

}
\description{
$ERG or token balance history of an address.
}
\details{
Sorts by oldest first with a limit of 10,000 rows.
}
