% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergowatch.R
\name{contractsCount}
\alias{contractsCount}
\title{Get Contract Address Count}
\usage{
contractsCount(bal_ge = 0, bal_lt = NULL, token_id = NULL)
}
\arguments{
\item{bal_ge}{Only count contract addresses with balance greater or equal to bal_ge. Default at 0.}

\item{bal_lt}{Only count contract addresses with balance lower than bal_lt.}

\item{token_id}{Token ID.}
}
\value{
A tibble with total count of contract addresses.
}
\description{
Current contract addresses count.
}
\examples{
contractsCount()
}
