% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergodex.R
\name{ed_allPoolStats}
\alias{ed_allPoolStats}
\title{All Pool Stats}
\usage{
ed_allPoolStats(from = NULL, to = NULL)
}
\arguments{
\item{from}{Window lower bound (UNIX timestamp millis).}

\item{to}{Window upper bound (UNIX timestamp millis).}
}
\value{
A tibble with all liquidity pool statistics.
}
\description{
Get statistics on all liquidity pools.
}
\examples{
ed_allPoolStats()

start <- 1648796400000
end <- 1648926788038
ed_allPoolStats(from = start, to = end)
}
