% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergodex.R
\name{ed_platformStats}
\alias{ed_platformStats}
\title{Platform Statistics}
\usage{
ed_platformStats()
}
\value{
A tibble with platform statistics for volume and TVL.
}
\description{
Get statistics on whole AMM.
}
\examples{
ed_platformStats()
}
