% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.R
\name{ergodexPlatformStats}
\alias{ergodexPlatformStats}
\title{Platform Statistics}
\usage{
ergodexPlatformStats()
}
\value{
A tibble with platform statistics for volume and TVL.
}
\description{
Get statistics on whole AMM.
}
\examples{
ergodexPlatformStats()
}
