% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergowatch.R
\name{ew_addressesBalanceHistory}
\alias{ew_addressesBalanceHistory}
\title{Address Balance History}
\usage{
ew_addressesBalanceHistory(address, token_id = NULL)
}
\arguments{
\item{address}{Required: the P2PK address.}

\item{token_id}{Token ID.}
}
\value{
A tibble with requested address, block height, timestamp/datetime, and balance.
}
\description{
$ERG or token balance history of an address.
}
\details{
Sorts by oldest first with a limit of 10,000 rows.
}
\examples{
bearwhale <- "9hyDXH72HoNTiG2pvxFQwxAhWBU8CrbvwtJDtnYoa4jfpaSk1d3"
ew_addressesBalanceHistory(bearwhale)

bearwhale <- "9hyDXH72HoNTiG2pvxFQwxAhWBU8CrbvwtJDtnYoa4jfpaSk1d3"
erdoge <- "3a7eb9b71a231b8df485c4d8b081726eda9876d2ce2d59bab5e5874c6cbba220"
ew_addressesBalanceHistory(bearwhale, token_id = erdoge)
}
