% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergowatch.R
\name{ew_addressesTags}
\alias{ew_addressesTags}
\title{Address Tags}
\usage{
ew_addressesTags(address)
}
\arguments{
\item{address}{Required: the P2PK or P2S address.}
}
\value{
A tibble with the tags associated to a given address.
}
\description{
Returns all tags associated for a given address (e.g. exchange address, known contract, etc.)
}
\examples{
coinex <- "9fPiW45mZwoTxSwTLLXaZcdekqi72emebENmScyTGsjryzrntUe"
ew_addressesTags(coinex)
}
