% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergowatch.R
\name{ew_contractsSupply}
\alias{ew_contractsSupply}
\title{Supply In Contracts}
\usage{
ew_contractsSupply(token_id = NULL)
}
\arguments{
\item{token_id}{Token ID.}
}
\value{
A tibble with total value supplied in contracts.
}
\description{
Supply in contracts.
}
\examples{
ew_contractsSupply()

migoreng <- "0779ec04f2fae64e87418a1ad917639d4668f78484f45df962b0dec14a2591d2"
ew_contractsSupply(migoreng)
}
