% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergowatch.R
\name{ew_exchanges}
\alias{ew_exchanges}
\title{List Tracked Exchanges}
\usage{
ew_exchanges()
}
\value{
A tibble with the ID's for exchanges being tracked on ErgoWatch.
}
\description{
List ID's of tracked exchanges.
}
\examples{
ew_exchanges()
}
