% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergowatch.R
\name{ew_exchangesSupply}
\alias{ew_exchangesSupply}
\title{Exchange Supply History}
\usage{
ew_exchangesSupply(exchange, since = NULL, limit = NULL)
}
\arguments{
\item{exchange}{Required: Exchange ID (use ew_exchanges() function).}

\item{since}{Height or timestamp of first record.}

\item{limit}{Maximum of 10,000.}
}
\value{
A tibble with the ID's for exchanges being tracked on ErgoWatch.
}
\description{
Supply on known main and deposit addresses.
}
\examples{
ew_exchangesSupply(exchange = "kucoin")

Note: the "since" and "limit" parameters MUST be used together.
ew_exchangesSupply(exchange = "gate", since = 1, limit = 100)
}
