% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergowatch.R
\name{ew_p2pkRanking}
\alias{ew_p2pkRanking}
\title{P2PK Address Rank}
\usage{
ew_p2pkRanking(address)
}
\arguments{
\item{address}{Required: the P2PK address.}
}
\value{
A tibble with the target address, rank, balance, and the next/previous addresses as well.
}
\description{
Get the rank of a P2PK address by current balance.
}
\examples{
bearwhale <- "9hyDXH72HoNTiG2pvxFQwxAhWBU8CrbvwtJDtnYoa4jfpaSk1d3"
ew_p2pkRanking(bearwhale)
}
