% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergowatch.R
\name{ew_richlistsBalance}
\alias{ew_richlistsBalance}
\title{Rich List}
\usage{
ew_richlistsBalance(token_id = NULL, limit = NULL)
}
\arguments{
\item{token_id}{Token ID.}

\item{limit}{Default value of 100 with maximum of 10,000.}
}
\value{
A tibble with the top addresses and their respective values for holding $ERG or a given token ID.
}
\description{
Get addresses with largest balance.
}
\examples{
ew_richlistsBalance()

comet <- "0cd8c9f416e5b1ca9f986a7f10a84191dfb85941619e49e53c0dc30ebf83324b"
ew_richlistsBalance(comet)
}
