% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergowatch.R
\name{ew_tokensDetail}
\alias{ew_tokensDetail}
\title{Token Details}
\usage{
ew_tokensDetail(token_id)
}
\arguments{
\item{token_id}{Required: Token ID.}
}
\value{
A tibble with token detail information: name, description, emission amount, decimals, and standard.
}
\description{
Details for a specific token.
}
\examples{
migoreng <- "0779ec04f2fae64e87418a1ad917639d4668f78484f45df962b0dec14a2591d2"
ew_tokensDetail(migoreng)
}
