% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2pk.R
\name{p2pkCount}
\alias{p2pkCount}
\title{Number of P2PK Addresses}
\usage{
p2pkCount(bal_ge = 0, bal_lt = NULL, token_id = NULL)
}
\arguments{
\item{bal_ge}{Only count contract addresses with balance greater or equal to bal_ge. Default at 0.}

\item{bal_lt}{Only count contract addresses with balance lower than bal_lt.}

\item{token_id}{Token ID.}
}
\value{
A tibble with total count of P2PK addresses.
}
\description{
Current P2PK addresses count.
}
\examples{
Total Ergo wallets
p2pkCount()

Wallets holding 1 million or more $ERG
p2pkCount(bal_ge = 1000000)

Count of wallets holding 1 billion or more $migoreng
migoreng <- "0779ec04f2fae64e87418a1ad917639d4668f78484f45df962b0dec14a2591d2"
p2pkCount(bal_ge = 1000000000, token_id = migoreng)
}
