% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{tokensSupply}
\alias{tokensSupply}
\title{Token Supply}
\usage{
tokensSupply(token_id)
}
\arguments{
\item{token_id}{Token ID.}
}
\value{
A tibble with total, circulating, and burned supply for a specific token.
}
\description{
Emitted, circulating, and burned token supply.
}
\examples{
migoreng <- "0779ec04f2fae64e87418a1ad917639d4668f78484f45df962b0dec14a2591d2"
tokensSupply(migoreng)
}
