% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addresses.R
\name{addressesBalanceTimestamp}
\alias{addressesBalanceTimestamp}
\title{Address Balance At Timestamp}
\usage{
addressesBalanceTimestamp(address, timestamp, token_id = NULL)
}
\arguments{
\item{address}{Required: the P2PK address.}

\item{timestamp}{Required: timestamp.}

\item{token_id}{Token ID.}
}
\value{

}
\description{
$ERG or token balance of an address at a specific timestamp.
}
