% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{p2pkRanking}
\alias{p2pkRanking}
\title{P2PK Address Rank}
\usage{
p2pkRanking(address)
}
\arguments{
\item{address}{Required: the P2PK address.}
}
\value{
work in progress.
}
\description{
Get the rank of a P2PK address by current balance.
}
\details{
Includes next and previous addresses as well.
}
\examples{
bearwhale <- "9hyDXH72HoNTiG2pvxFQwxAhWBU8CrbvwtJDtnYoa4jfpaSk1d3"
p2pkRanking(bearwhale)
}
