% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{tokensSupply}
\alias{tokensSupply}
\title{Token Supply}
\usage{
tokensSupply(token_id)
}
\arguments{
\item{token_id}{The token ID on the Ergo blockchain.}
}
\value{
A tibble with total, circulating, and burned supply for a specific token.
}
\description{
Emitted, circulating, and burned token supply.
}
\examples{

}
