% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{tokensSupply}
\alias{tokensSupply}
\title{Token Supply}
\usage{
tokensSupply(token_id)
}
\arguments{
\item{token_id}{Required: Token ID.}
}
\value{
A tibble with token supply breakdown: emitted, in P2PK addresses, in contracts, and burned. Emitted is sum of other three.
}
\description{
Emitted, circulating (P2PK addresses vs contracts), and burned token supply.
}
\examples{
migoreng <- "0779ec04f2fae64e87418a1ad917639d4668f78484f45df962b0dec14a2591d2"
tokensSupply(migoreng)
}
